requirements.txt
uvicorn==0.23.2
fastapi==0.101.0
minio==7.1.16
pillow